ci: clear the lint gates for #8872
The `lint` job failed on four gates that the PR's own changes tripped:

- changelog: add the `changelog.d/8872-*` fragment for the crates/ changes.
- file size: `array/indexing.rs` reached 2,024 lines after the resolved-store
  work; move the transactional `js_array_numeric_range_add*` kernel (a block
  with no raw-handle or address-classification debt, so no per-module ratchet
  ceiling moves) into `array/numeric_range.rs`.
- local-binding-type audit: classify the synthetic `arguments.length` marker
  read in `property_get.rs::lower` (runtime-validated: the marker type exists
  only in direct-call-only clones whose caller materialized the count).
- GC store-site inventory: register `store_array_slot_resolved` as a
  chain-verified discharge helper for the three BARRIERED markers that now
  lean on it, mark its own resolved slot write, and pin the second `apush`
  codegen marker (the unconditional element store inside
  `emit_dynamic_pointer_push_store`, barriered by the same stem) with the
  self-test tree updated to match.

Every step of the lint job was replayed locally, including the raw-handle and
unrooted-local ratchets against the merge base d354443.
